--- pyproject.toml
@@ -1,27 +1,0 @@
-[project]
-name = "print_it"
-version = "0.1.0"
-description = "A printing utility"
-requires-python = ">=3.10"
-dependencies = [
-    "streamlit>=1.26.0",
-    "Pillow>=10.0.0",
-    "pyusb>=1.2.1",
-    "qrcode>=7.4.2",
-    "requests>=2.31.0",
-    "brother-ql-inventree @ git+https://github.com/matmair/brother_ql-inventree.git@master",
-]
-
-[project.urls]
-repository = "https://github.com/yair/print_it"
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.metadata]
-allow-direct-references = true
-
-
-[tool.hatch.build.targets.wheel]
-packages = ["."]
